chore: refactor constants & interfaces file location
