+++ hooks/useVerifyCandyMachineV2.tsx
@@ -2,8 +2,8 @@ import { AnchorProvider, BN, Program } from '@project-serum/anchor'
 import { useAnchorWallet, useConnection } from '@solana/wallet-adapter-react'
 import { PublicKey } from '@solana/web3.js'
 import { ICache, saveCache } from 'lib/cache'
-import { CANDY_MACHINE_PROGRAM_V2_ID, CONFIG_ARRAY_START_V2, CONFIG_LINE_SIZE_V2 } from 'lib/constants'
-import { Account } from 'lib/types'
+import { CANDY_MACHINE_PROGRAM_V2_ID, CONFIG_ARRAY_START_V2, CONFIG_LINE_SIZE_V2 } from 'lib/candy-machine/constants'
+import { Account } from 'lib/candy-machine/types'
 import { getTextFromUTF8Array, shardArray } from 'lib/utils'
 import { useState } from 'react'
