+++ lib/cache/constants.ts
@@ -1,0 +1,1 @@
+export const CACHE_PATH = './.cache'
