+++ pages/[id].tsx
@@ -2,9 +2,9 @@ import { AnchorProvider, BN, Program } from '@project-serum/anchor'
 import { useAnchorWallet, useConnection } from '@solana/wallet-adapter-react'
 import { Connection, PublicKey } from '@solana/web3.js'
 import { Spinner, Title, UpdateCreateCandyMachineForm } from 'components'
-import { CANDY_MACHINE_PROGRAM_V2_ID } from 'lib/constants'
-import { IFetchedCandyMachineConfig } from 'lib/interfaces'
-import { Account } from 'lib/types'
+import { CANDY_MACHINE_PROGRAM_V2_ID } from 'lib/candy-machine/constants'
+import { IFetchedCandyMachineConfig } from 'lib/candy-machine/interfaces'
+import { Account } from 'lib/candy-machine/types'
 import type { NextPage } from 'next'
 import Head from 'next/head'
 import { useRouter } from 'next/router'
